feat(benchmark): tryin' to get the render time for the compoents.
