+++ packages/duck-benchmark/src/utils/compile-benchmark/compile-benchmark.types.ts
@@ -1,0 +1,6 @@
+import { FolderInfo } from '../list-files'
+
+export type CompileBenchmarkParams = {
+  folders: FolderInfo[]
+  visited?: Set<string>
+}
