+++ packages/duck-benchmark/src/utils/compile-benchmark/index.ts
@@ -1,0 +1,3 @@
+export * from './compile-benchmark'
+export * from './compile-benchmark.libs'
+export * from './compile-benchmark.constants'
