+++ packages/duck-benchmark/src/utils/list-files/index.ts
@@ -1,2 +1,3 @@
 export * from './list-files'
 export * from './list-files.types'
+export * from './list-files.dto'
