+++ packages/duck-benchmark/src/utils/list-files/list-files.types.ts
@@ -1,5 +1,8 @@
+import { Ora } from 'ora'
+
 export interface ListFilesOptions {
   cwds: string[]
   depth?: number
   filter?: string[]
+  spinner: Ora
 }
