+++ packages/duck-benchmark/tsconfig.json
@@ -18,7 +18,7 @@
         "resolveJsonModule": true,
         "baseUrl": ".",
         "paths": {
-            "@/*": [
+            "~/*": [
                 "./*"
             ]
         }
